Migrate from GitHub + split to mainline/stable
This is a 2-in-1 deal. There was a takedown of the suyu GitLab repos,
and the team has since moved to their own personal Forgejo instance
(pretty much a Gittea instance). I originally was going  to add the
mainline/stable split once I saw that there was tagging done. Since both
had to be done at once, here we are!